+++ lib/builtin/canvas/canvas.dart
@@ -1,1 +1,2 @@
+export 'src/canvas_pane_host.dart';
 export 'src/extension.dart';